Improve primitive
Use an option for primitive without a material and 16 bits for indices.

+++ src/rustspot/gfx.rs
@@ -239,7 +239,7 @@ impl Ebo {
         unsafe { gl::BindBuffer(gl::ELEMENT_ARRAY_BUFFER, self.handle) };
     }
 
-    pub fn upload(&mut self, indices: &Vec<u32>) {
+    pub fn upload(&mut self, indices: &Vec<u16>) {
         self.bind();
         unsafe {
             gl::BufferData(
